Round 2 of wx3 changes on OSX
Main change is that VST GUI support is now integrate with new Cocoa
views.  Direct support for VST Cocoa views (via Cockos extensions:
http://www.reaper.fm/sdk/vst/vst_ext.php) has been added.

+++ src/prefs/TracksPrefs.cpp
@@ -36,7 +36,7 @@ TracksPrefs::TracksPrefs(wxWindow * parent)
    // Bugs 1043, 1044
    // First rewrite legacy preferences
    gPrefs->Write(wxT("/GUI/DefaultViewModeNew"),
-      WaveTrack::FindDefaultViewMode());
+      (int) WaveTrack::FindDefaultViewMode());
 
    Populate();
 }
